ci: fix workflows and resolve all golangci-lint findings
WORKFLOW FIXES:
- Migrate .golangci.yml from broken v1-schema config to real v2 format
  (linters.settings, linters.exclusions.rules, run.timeout)
- Reorder steps: checkout before setup-go so cache can find go.sum
- Run tests before lint so coverage.out is always created
- Drop redundant actions/cache (setup-go handles it natively)
- Upgrade all GitHub Actions to Node 24:
  - checkout v4→v5, setup-go v5→v6
  - golangci-lint-action v6→v9, upload-artifact v4→v7
  - action-gh-release v2→v3

CODE FIXES (139 findings):
- Add doc comments to all exported types/funcs/vars (100 revive issues)
- Handle all unchecked errors with _ = <expr> (20 errcheck issues)
- Fix qqwry bounds check: uint64 to prevent overflow + end+7 wrap bug
- Remove unused Reader.seek method (wry package)
- Type explicit const groups: Format/Type in internal/db/type.go
- Rename stuttering types: CDNResult→Result, EntityType→Type, etc.
- Tighten file permissions: dir 0755→0750, WriteFile 0644→0600
- Annotate legitimate G304 file-inclusion uses (user-configured DB paths)

Verified: lint 0 issues, build OK, tests pass (race+coverage)

+++ pkg/qqwry/qqwry.go
@@ -14,6 +14,7 @@ import (
 	"github.com/zu1k/nali/pkg/wry"
 )
 
+// DownloadUrls are the mirror URLs the QQwry database is fetched from.
 var DownloadUrls = []string{
 	"https://github.com/metowolf/qqwry.dat/releases/latest/download/qqwry.dat",
 	// Other repo:
@@ -22,6 +23,7 @@ var DownloadUrls = []string{
 	// https://github.com/metowolf/qqwry.dat
 }
 
+// QQwry is a QQwry (纯真) IPv4 geolocation database reader.
 type QQwry struct {
 	wry.IPDB[uint32]
 }
@@ -63,6 +65,7 @@ func NewQQwry(filePath string) (*QQwry, error) {
 	}, nil
 }
 
+// Find looks up query and returns its location, or an error for non-IPv4 input.
 func (db QQwry) Find(query string) (result fmt.Stringer, err error) {
 	ip := net.ParseIP(query)
 	if ip == nil {
@@ -87,10 +90,12 @@ func (db QQwry) Find(query string) (result fmt.Stringer, err error) {
 	return reader.Result.DecodeGBK(), nil
 }
 
+// Name returns the database name.
 func (db QQwry) Name() string {
 	return "qqwry"
 }
 
+// CheckFile reports whether data looks like a valid QQwry database.
 func CheckFile(data []byte) bool {
 	if len(data) < 8 {
 		return false
@@ -100,7 +105,8 @@ func CheckFile(data []byte) bool {
 	start := binary.LittleEndian.Uint32(header[:4])
 	end := binary.LittleEndian.Uint32(header[4:])
 
-	if start >= end || uint32(len(data)) < end+7 {
+	// Compare in uint64 so neither the len conversion nor end+7 can overflow.
+	if start >= end || uint64(len(data)) < uint64(end)+7 {
 		return false
 	}
